project-5-hello-langchain/requirements.txt
<<<<<<< HEAD
langchain
langchain-google-genai
python-dotenv
langchain-community
=======
openai
python-dotenv
gradio
>>>>>>> cc9230b4d031a1619f9a33586f7fa5761d557fa3
